Add keep original file option

+++ cbr2cbz.rb
@@ -1,5 +1,21 @@
 # encoding: utf-8
 
+require 'optparse'
+
+options = {}
+OptionParser.new do |opts|
+  opts.banner = "Usage: cbr2cbz [options] file ..."
+
+  opts.on("-k", "--keep", "Keep original file") do |k|
+    options[:keep] = k
+  end
+
+  opts.on_tail("-h", "--help", "Show this message") do
+    puts opts
+    exit
+  end
+end.parse!
+
 require 'fileutils'
 require 'zip/zip'
 require 'zip/zipfilesystem'
@@ -7,18 +23,18 @@ require 'zip/zipfilesystem'
 ARGV.each do |arg|
   files = File.directory?(arg) ? Dir.glob(arg + '/**') : [arg]
   files.each do |filename|
-    if File.extname(filename) == '.cbr'
+    if File.extname(filename).downcase == '.cbr'
       FileUtils.cd(File.dirname(filename)) do
 
-        # Unrar CBR
         filename_without_ext = File.basename(filename, '.*')
 
+        # Unrar CBR
         `unrar e "#{filename}" "#{filename_without_ext}"/`
 
         # Create CBZ
-        archive = filename_without_ext + '.cbz'
+        new_file = filename_without_ext + '.cbz'
 
-        Zip::ZipFile.open(archive, 'w') do |zipfile|
+        Zip::ZipFile.open(new_file, 'w') do |zipfile|
           Dir[filename_without_ext + "/*"].each do |file|
             zipfile.add(File.basename(file),file)
           end
@@ -26,7 +42,7 @@ ARGV.each do |arg|
 
         # Clean up
         FileUtils.rm_rf(filename_without_ext)
-        FileUtils.rm(filename)
+        FileUtils.rm(filename) unless options[:keep]
       end
     end
   end
